Move mempool tests to separate file
